Moved IntegrationTest base into testutils package
