+++ src/test/kotlin/uk/gov/justice/digital/assessments/controller/PingControllerTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient
 import org.springframework.test.web.reactive.server.expectBody
 import uk.gov.justice.digital.assessments.controllers.PingController
+import uk.gov.justice.digital.assessments.testutils.IntegrationTest
 
 @AutoConfigureWebTestClient
 class PingControllerTest : IntegrationTest() {
