+++ src/test/kotlin/uk/gov/justice/digital/assessments/controller/QuestionControllerTest.kt
@@ -11,6 +11,7 @@ import uk.gov.justice.digital.assessments.api.GroupQuestionDto
 import uk.gov.justice.digital.assessments.api.GroupSummaryDto
 import uk.gov.justice.digital.assessments.api.GroupWithContentsDto
 import uk.gov.justice.digital.assessments.api.QuestionSchemaDto
+import uk.gov.justice.digital.assessments.testutils.IntegrationTest
 import java.util.*
 
 @SqlGroup(
