+++ src/test/kotlin/uk/gov/justice/digital/assessments/repositories/QuestionSchemaRepositoryTest.kt
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.jdbc.Sql
 import org.springframework.test.context.jdbc.SqlConfig
 import org.springframework.test.context.jdbc.SqlGroup
-import uk.gov.justice.digital.assessments.controller.IntegrationTest
+import uk.gov.justice.digital.assessments.testutils.IntegrationTest
 import uk.gov.justice.digital.assessments.jpa.repositories.QuestionSchemaRepository
 import java.util.*
